+++ src/i18n/template.ts
@@ -4,10 +4,10 @@ import en from './en.js';
 
 const overrides = {
 
-  // TODO copy strings from en.js to here and translate
+  // copy strings from en.js to here and translate
 
 };
 
-const TODO_REPLACE_THIS_WITH_ISO_CODE = merge({}, en, overrides);
+const REPLACE_THIS_WITH_ISO_CODE = merge({}, en, overrides);
 
-export default TODO_REPLACE_THIS_WITH_ISO_CODE;
+export default REPLACE_THIS_WITH_ISO_CODE;